move logdistributor into code
